Created Empty Files for The Project
Created three empty files (index.html, script.js, and styles.css).

index.html will be for the html of the project, to display what the user
is able to see and interact with.

styles.css will provide the styling for the sign-up for.

script.js will provide the javascript, making the webpage interactable.