Added Javascript From Previous Sign Up From
Added javascript from the previous sign up form.

This validates the inputs and displays the pop-up windows upon
successful registration.

Additionally, the script allows users to login through the login pop-up
window and displays the welcome back pop-up upon successful login. The
script only verifies that a username and password has been entered for
logging in, users and passwords are not stored in a database.